requirements.txt
functions-framework==3.*
google-cloud-storage
torchvision
torch
opencv-python-headless